render: split render.go into focused files + quality fixes
Refactor
--------
Break render/render.go (2312 lines) into 11 focused files, all under
600 lines and none changing behaviour. Top-level map:

- render.go          (Render, buildReport, RenderOptions) ~100 lines
- assets.go          (embedded FS + JSON loaders + variable classifier)
- collection.go      (Collection → Report section builders)
- concat.go          (per-collector time-axis concatenators)
- navigation.go      (flat NavEntry builder + nav grouping)
- view.go            (reportView + every sub-view struct)
- build_view.go      (Report → reportView flattening)
- payloads.go        (chart JSON payload builders)
- summaries.go       (chart summary-row builders + helpers)
- dedup.go           (snapshot dedup signatures and range notes)
- innodb.go          (aggregateInnoDBMetrics + semaphore breakdown)
- advisor_view.go    (findings → findingView)

Each top-level declaration moved verbatim with its surrounding
godoc. `make test` and `make vet` green — the rendered HTML is
byte-identical to pre-refactor output (Constitution Principle IV).

New tests (Task 2)
------------------
- render/innodb_test.go covers aggregateInnoDBMetrics,
  buildAHIMetric (including the 31.0% hit-ratio canonical case and
  the zero-activity fallback), attachSemaphoreBreakdown (peak +
  window-total math), and the isMysqldCommand mariadbd match.
- render/dedup_test.go covers snapshotSignature (volatile vars
  ignored, case-insensitive), the [A,A,B,B] collapse preserving
  first panel and flagging only the real diff on the second,
  formatRangeNote/parseRangeNote roundtrip, and extendRange on
  an unseeded note.
- render/os_test.go TestOSSubviewAnchors now also asserts the
  sub-os-meminfo anchor + nav href.

Quality fixes (Task 3)
----------------------
- MED #2   isMysqldCommand matches mariadbd as well as mysqld.
- MED #3   parse/meminfo.go emits a SeverityWarning diagnostic per
           declared series that never appeared in any sample.
- LOW #4   Replace ins[:0] aliasing in concatVmstat/concatMeminfo
           (and by consistency concatIostat/concatTop) with a
           fresh make-with-cap so the filter loop cannot overwrite
           caller memory.
- LOW #5   Delete dead innoDBFloatMetric.
- LOW #6   parse/innodbstatus emits a SeverityWarning when
           sum(SemaphoreSites[*].WaitCount) != threadWaited so
           a silent miscount surfaces in stderr.
- LOW #7   parse/innodbstatus flushes any pending semaphore record
           on section change so the last thread-without-partner
           is not lost.
- LOW #10  innoDBMetricView.AHIHasHashTable bool replaces the
           brittle template-side string compare against the dash
           placeholder. Templates now gate on the flag.
- LOW #11  extendRange treats a (0,0) parseRangeNote result as
           "no range seeded" and anchors on snapNum.
- NIT #27  snapshotSignature is computed once in buildReport and
           shared with buildNavigation + buildView, eliminating
           the duplicate pass.
- NIT #43  concatVmstat/concatMeminfo share a new
           mergeMetricSeriesByName helper with the per-caller
           Unit policy ("mixed"-promoted for vmstat, fixed "GB"
           for meminfo) preserved verbatim.
- NIT #44  model.SortSemaphoreSites centralises the descending-by-
           WaitCount-with-File/Line/Mutex tie-break ordering used
           by both parse/innodbstatus and render/attachSemaphore-
           Breakdown.

Skipped
-------
LOW #39 (CSS comment on iris clip origin) and LOW #40 (add
::marker {display:none} for Firefox) are deferred. Both would
modify the embedded app.css bytes, which embed directly into the
rendered HTML and therefore would violate the byte-identical
output constraint (Constitution Principle IV) for this refactor
pass. They can land in a follow-up that explicitly regenerates
the goldens.
