+++ parse/iostat_test.go
@@ -166,4 +166,3 @@ func TestIostatPartialRecovery(t *testing.T) {
 			truncatedSource, len(diags), diags)
 	}
 }
-
